+++ Cargo.toml
@@ -21,6 +21,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 thyllore-math-core = { path = "crates/thyllore-math-core" }
 thyllore-model-core = { path = "crates/thyllore-model-core" }
+thyllore-anim-core = { path = "crates/thyllore-anim-core" }
 anyhow = "1"
 log = "0.4"
 cgmath = "0.18"
