+++ crates/thyllore-anim-core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "thyllore-anim-core"
+version = "0.0.1"
+edition = "2021"
+license = "Apache-2.0"
+authors = ["kodai731"]
+description = "Layer 2 dynamic animation evaluation for ThylloreAnimation (Clip, Pose, Constraint, SpringBone, editable) — pure domain, no ECS dependency"
+repository = "https://github.com/kodai731/Thyllore-Animation"
+
+[dependencies]
+thyllore-math-core = { path = "../thyllore-math-core" }
+thyllore-model-core = { path = "../thyllore-model-core" }
+cgmath = "0.18"
+serde = { version = "1", features = ["derive"] }
+ron = "0.8"
+anyhow = "1"
+
+[dev-dependencies]
+serde_json = "1"
