+++ crates/thyllore-anim-core/src/editable/components/clip_instance.rs
