+++ crates/thyllore-anim-core/src/editable/components/mod.rs
@@ -4,6 +4,8 @@ pub mod clip_group;
 pub mod clip_instance;
 pub mod curve;
 pub mod keyframe;
+pub mod keyframe_copy;
 pub mod mirror;
+pub mod snap_settings;
 pub mod source_clip;
 pub mod track;
