+++ crates/thyllore-anim-core/src/lib.rs
@@ -1,10 +1,12 @@
+mod clip;
 mod constraint;
-mod data;
 pub mod editable;
 pub mod keyframe_search;
 mod pose;
 pub mod spring_bone;
 
+pub use clip::*;
 pub use constraint::*;
-pub use data::*;
 pub use pose::*;
+
+pub use thyllore_model_core::{Bone, BoneId, Skeleton, SkeletonId, SkinData};
