+++ crates/thyllore-anim-core/src/spring_bone/chain.rs
@@ -1,4 +1,4 @@
-use crate::animation::{BoneId, Skeleton};
+use crate::{BoneId, Skeleton};
 
 pub fn build_evaluation_order(
     chains: &[(Vec<BoneId>,)],
