+++ src/animation.rs
@@ -1,0 +1,1 @@
+pub use thyllore_anim_core::*;
